[Trusted Types] Worker constructor URLs are TrustedScriptURLs
Make URLs used to construct workers TrustedScriptURLs.

Minor rework of how to get the execution context in the generated bindings,
because the old method assumed presence of an "impl" object, whichfor is
only created at the end in case of constructors. So insetead, implement a
method with which we can always obtain the execution context from the
v8::FunctionCallbackInfo.

Bug: 1059684
Change-Id: I00e3e52cad5c3736105289a9fda1eeca15fd6f62