SVG: Fix clipPath for positioned <foreignObject>
In this scenario, the x & y attributes are baked into both the the
content-rect of the <foreignObject> (as passed to the clipper), *and*
the transform.  As a result, any x & y offsets were being double-applied
to a <clipPath> that used clipPathUnits="objectBoundingBox".

Our solution is to prevent the clipper from unnecessarily applying the
offset when calculating the the clip transform.  (To avoid side-effects,
we only make this change for SVG foreign objects.)

Bug: 1009222
Change-Id: Ifc009eb84d46d49c7840e2903d44a90502d6b51d