Let animations/transitions use the main cascade
In order to solve difficult problems such as crbug.com/552085,
interpolations must be applied using the same cascade as where the
MatchResult was analyzed. This means we'll skip applying the
interpolation if an !important declaration exists for the same
property.

This CL:

 - Moves the MatchResult out of ElementRuleCollector. MatchResult must
   now be passed to ApplyAnimatedStandardProperties, hence it must
   outlive ElementRuleCollector.
 - Lifts the on-stack StyleCascade up to StyleForElement, such that it
   can be used in both ApplyBaseComputedStyle and ApplyAnimated-
   StandardProperties.
 - Sets as flag on StyleResolverState when an interpolation effect is
   skipped during Apply. This means something was !important, and we
   can't use the base-computed-style optimization.
 - Clears the base-computed-style if the above flag is set.
 - Due to how the base computed style optimizations works when DCHECK
   is on, it was necessary to add Reset functions to StyleCascade and
   MatchResult (see explanation near MaybeResetCascade).

This is a web-facing change: !important declarations will now override
declarations in the animation origin. WPT for this was apparently non-
existent, so this CLs adds a few tests (that also pass in Firefox).

Bug: 552085, 985049
Change-Id: I538f73936e96e1bd974e0855b081e99bf947e6a2